requirements.txt
matplotlib==3.3.4
pandas==1.1.5
scipy==1.5.2
numpy==1.17.0
tensorflow_gpu==1.15.0
